requirements.txt
# NLP
jieba>=0.42.1
nltk>=3.8
cn2an==0.5.22
word2number>=1.1
roman-numbers>=1.0.2

# Document Parsing
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23
markdown>=3.5
beautifulsoup4>=4.12.0
PyMuPDF>=1.24.0
chardet>=5.2.0

# Embedding & Search
openai>=1.0.0
elasticsearch>=8.0.0
elasticsearch-dsl>=8.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# API
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9

# GraphRAG
networkx>=3.0
json-repair>=0.25.0

# Utils
tiktoken>=0.7.0
pyyaml>=6.0.1
xxhash>=3.4.0
